B-52830 - Upgrading compileJava build task to support gradle 1.6
 Under gradle 1.4 the standard way to get aspectj working in gradle was to overwrite the standard compileJava included in gradle ala:

 task compileJava(dependsOn: JavaPlugin.PROCESS_RESOURCES_TASK_NAME, overwrite: true)  {

However, in gradle 1.5 and subsequently 1.6, this method no longer worked. A previously working buid would suddenly get an error at runtime similar to:

org.springframework.beans.factory.BeanCreationException: Error creating bean with name 'validateAspect' defined in class path resource [app-config.xml]: No matching factory method found: factory method 'aspectOf'. Check that a method of the specified name exists and that it is static.

A few articles on the web described this issue at:
http://forums.gradle.org/gradle/topics/aspectj_plugin_does_not_work_with_gradle_1_5
and
http://forums.gradle.org/gradle/topics/replacing_tasks_overwrite_true_does_not_work_in_1_5

The suggested solution is to not override the entire task as previously, but simply the actions within the task.