[B-54660] - Fixing failing idmPropertiesTest due to inclusion of Vagrant config section
The vagrant config was based off the jenkins config. The idmPropertiesTest ignores the jenkins config so properties contained in the jenkins config that are not in the other properties files do not cause an error. However, when the vagrant config was added it was not excluded from the test so the additional properties copied from the Jenkins config caused errors. The fix was excluding the vagrant config is well. However, this overall test should be reworked as the current state does not appear to actually meet the expectation of the test.