[D-14952] - Fixes a defect in how the password is retrieved from the environment variable
Previous code would incorrectly test whether the Algorithm was set to determine if the password should be read from the password environment variable. The algorithm is set prior to that code running, so the password environment variable would never be read. While fixing, noticed there was a bug in 3rd party code that would cause a failure if the password env variable was set to an empty string (even if the system property was set to the correct value). Also noticed the provider name system property overrode the env variable, which is contrary to documentation.