[D-14829] - Modify subtree deletion to be single operation
As I dug into the fix, the scope of the changes has increased somewhat. Want to provide a more detailed update:

1. Subtree deletion
This was a relatively minor change to make with respect to production code. However, the QE and dev regression test impact is/was significantly larger since all code that deletes from CA eventually calls this single method.

2. Test Refactor
Addressing the resultant broken tests for this was pretty complex due to the structure of the code (superclass/subclass) and the fact that some method signatures specified the return value as a final concrete classes rather than interface which made for interesting mocking. I did some small refactors here to eliminate the myriad of hoops and spies that were required, and migrated tests of the superclass methods (in LdapRepository) to be in a test class specifically for the superclass. I only migrated those that I had to touch/fix for this story so LdapEndpointRepositoryTest still contains some tests for the superclass. These tests also needed to be in junit/java rather than spock/groovy due to mockito features required that are not supported in groovy.  Note - I did NOT look at all the other locations that ultimately call the subtree deletion to make sure they still work and/or actually have tests that use it. At this point I don't plan on verifying their test code coverage.

3. Concurrency issues in test cases
 I _finally_ figured out why my dev regression test for Item 1 was failing. It was a concurrency bug in an unrelated part of the app (DefaultAuthorizationService). The DefaultAuthorizationService class used static variables to "cache" various roles (cloudServiceAdminRole, cloudUserAdminRole, etc). We then have unit tests (DefaultAuthorizationServiceTestOld.java/DefaultAuthorizationServiceTest.groovy) that set these static variables to non real values to test some functionality and then never set them back. Integration tests that happen to run after these unit tests (lucky me…mine) depend on these variables to be valid (in order to get a token for the service admin user). Hence, KABOOM.

 While the simple "fix" would likely have been to just make the tests set the static variables back to their initial values at the end, I was concerned that we were using static variables for this in the first place.
  - This is a spring managed singleton so there's no performance gain to using a static
  - The routine to set this value is a post instantiation routine that is supposed to run after every instance is created and is not synchronized, so if we ever did create multiple instances we would have concurrency issues
  - The static setters/getters are not synchronized
  - As described earlier, our tests stomp on top of each other

  To resolve the issue I changed the static variables to be instance variables, modified the getters/setters accordingly, and updated test cases as appropriate.

4. Updates per dao-refactor: Added multiple test cases for genericldaprepository addObject/deleteObject