[B-56554] - Fixed slow test in LDAPConfigurationTest
- Fixed nohosts test that didn't actually test what it was supposed to (that no host provided) and was running extremely slow.

The test attempts to verify the prod code wraps an exception thrown by the unboundid if hosts were not provided against which to connect (assume meant ports). However, the test actually sent in hosts, but numeric ones, which caused unboundid to try to establish connections to these hosts (interpreted as IPs). Since the hosts are invalid the timeout would need to be hit before unboundid throws an exception. The prod code doesn't specify a timeout to unboundid so the default timeout was used (which is 60 seconds). Since 2 connections were attempted, this resulted in this simple test taking 2 minutes instead of being pretty much instananeous. The data was fixed so it would be interpreted as ports rather than hosts.

 Also made minor modifications to other tests to use standard 'fail' JUnit method when want to force a test failure.