[D-15312 & D-15313]: Addressed user-managers updateUser permissions for self and other usermanagers
These two defects were tied at the hip - requiring a change to the exact same line and adding test cases to the same classes, and modifying the same other test cases. Rather than issuing one pull request, waiting for it to be merged, and then issue a subsequent pull request I just combined the two since it was such a small change.