[B-62438] - Added functionality to request ldap logs for any API and retrieve log via REST service
Logging will be disabled at the end of the request. Requires requests to be sequential as logfile value is shared across requests.

REQUEST HEADERS
Adding extra headers to any cloud request will enable various types of logging:

(Required to enable logging)
X-CREATE-LDAP-LOG - will enable logging, and create a log file for this request (default: false)
X-LDAP-LOG-TOKEN - a service admin token, required to enable ldap logging

(Optional)
X-RETURN-LDAP-LOG - whether or not to replace the response body with the contents of the log file. When enabled, the response status code will always be set to 200. However, a custom header will contain the original response code returned from the service.

RESPONSE HEADERS
When logging is enabled some additional response headers will be provided:
X-LDAP-LOG-LOCATION - the location of the log file. Can use retrieval REST service to get
X-LDAP-LOG-SERVICE-STATUS-CODE - if X-RETURN-LDAP-LOG is set to true, this is the original status code the service returned.