[B58490] - Added unit test for phone number generator
When the libphonenumber library is upgraded there is a chance that previously "valid" area codes are no longer considered valid. This test will ensure when this happens we verify the test util that it's generated number are still considered valid.

Note - Duo security uses this same library to validate phone numbers