[B79471] - Fixed issue with UUID impersonator impersonating AE User
When User A impersonates User B two tokens are in play - the impersonated token provided to User A, and the linked token issued against User B (to which the impersonated token links). Both these tokens should be of the same format, and determined by the impersonator.

There was a bug whereby if the impersonator used UUID tokens, but the impersonated user used AE tokens, the impersonated token would be a UUID token, but the linked user token would be an AE token. This has been fixed.