[CIDMDEV-5333] - Fix create IDP validation failures when providing empty string arguments
Fixed multiple validation failures when providing empty strings as arguments to description, approvedDomainGroup, approvedDomainIds.

- Providing empty string description will be ignore (stored as null) - 201 Result
- When provide non-null (or empty string) values for BOTH approvedDomainGroup and approvedDomainId, will get 400 error. Caller can only provide one or the other, not both - even if they set one to empty string.
- When provide non-null value for either approvedDomainGroup or approvedDomainIds for Racker IDP, get 400 error

Reason behind militant restriction on checking just for presence of the attribute (as opposed to only failing if user actually provides a value) is that trying to autocorrect the user input opens up a slew of edge cases that would need to be tested and handled. By restricting user to only providing one and only one of either approvedDomainGroup or approvedDomainId it's a much simpler service.