[CID-177] - Remove CXF dependency
We include CXF as a dependency, even though we don’t actually use CXF to provide our web services (we use jersey). However, there are various places in the code that use some of the CXF utilities. The uses are trivial and can easily be removed. This story does just that. Eliminates CXF as a dependency and porting current uses to comparable things.

The woodstox dependency that was pulled in transitively by CXF is required for XML processing so that was added explicitly.