[CID-594] - Made jcl-over-slf4j and woodstox dependencies explicit
The jcl dependency was a compile time transitive dependency from the MFA SDK. However, the opensaml upgrade required a newer version. The explicit dependency in the main build.gradle file list this as a runtime dependency, not a compile time. To be consistent with the MFA SDK, switched the explicit dependency to compile time, and upgraded to new version.

The woodstox change was due to the opensaml having a transitive dependency on a higher version. We explicitly listed a lower version so simply upgraded our explicit listing to be the newer version.

In order to easily compare dependency changes I also included a simple task to flatten and sort the dependencies list that gradle produces.