[CID-491] - Initial implementation to return appropriate roles on apply_rcn_roles
This is a basic implementation simply to get the overall architecture there. It's not overly concerned with optimization and focuses on just returning the right tenant roles. Includes a few tests for a subset of scenarios, but needs significant expansion. The primary goal was to get the roles available in order to unblock other dependent stories.