[CID-967] - Send final user update feed event when role changes complete
This just sends one final feed event for each user after all the roles have been updated on the users appropriately.
A feed event is also sent for each user when the new identity classification role is added to the user. This means 2 events are sent per user. The first event will show the new role getting added. The second will show the final state of the user's roles.