[CID-1017] - Switched impersonator to dynamically generated one to avoid test failures
The existing tests verifying impersonation tokens hardcoded the expected domain for the base service admin included in docker. However, other tests change the domain id of this user (they shouldn't, but they do). Since the tests already create an identity admin capable of impersonating, this change switches to using this user. This avoids an error that occurs if the Cloud20DomainIntegrationTests happen to run prior to these tests.