[CID-990] - Moved common test code to unbroken-dome based sourceset (#1597)
* [CID-990] - Moved common test code to unbroken-dome based sourceset

Intellij does not find usages of code marked as "test" within code marked as "production". This leads to incorrect assumptions about the use of the common test code within our commit and nonCommit tests. It also causes errors when trying to use intellij refactoring.

This moves the common test code to be managed by its own unbroken-dome sourceset similar to the commit/nonCommit tests. Intellij will then be able to find usages of the common code across the the tests.

Actual tests should NOT be put in this sourceset - only utility/shared code. Tests should only be put in the appropriate commit/nonCommit location.