[CID-1109] - Added javadoc for tenant services (#1612)
To help determine impact of applying user group membership to users when determining the user's roles we need a better understanding of the dozens of services within this class and how they are used. The majority had no javadoc documenting their use. To help with the next step of actually going through and following the threads of execution and subsequently impact, I needed to document all the various services.