[CID-1252] - Added ability to securely hash protected attributes (#1692)
* [CID-1252] - Added ability to securely hash protected attributes

Added option to create an HMAC (SHA1) of a subset of attributes prior to sending to New Relic in order to obfuscate it and remove any potential PII or sensitive information.  Uses an HMAC rather than simply a hash (or salted hash) in order to prevent dictionary attacked and for meaningful comparison across multiple days worth of data.

The HMAC (160 bits) is Base64 encoded prior to sending to New Relic to make it readable. This results in ~28 character string for all secured attributes.

- **feature.enable.secure.new.relic.api.resource.attributes**: (boolean; default = true). Whether to enable the feature to secure custom attributes. Only relevant is sending New Relic data itself is enabled.
- **new.relic.secured.api.resource.key**: (String; default=aRandomKey). The key to use in HMAC generation
- **new.relic.secured.api.resource.attributes**: (comma delimited list of strings; default=callerUsername). The set of attributes to secure prior to sending to New Relic. The value '*' can be used to secure *all* custom attributes, but that is NOT recommended.