CID-1585: Updating effective role calc to support users with invalid domains (#2072)
The switch to using the “new” standard for calculating a user’s roles had explicit validations that the user:
1. Has a domainId set on the user
2. The domainId actually resolves to an existing domain

For the record - both states are invalid data states. That _shouldn’t_ occur and, if it does, results are sorta unpredictable across the services considering we require all users to have a domain. We can’t account for all the potential bad data states and not fail. That said, you’ll prob recall earlier issues where we ran in to domains not existing and we wrote explicit logic to _not_ fail authentication for these users - even though it’s an invalid state. We even have a few nonC tests that explicitly verify this is allowed (hence the failing tests).

Long term, we need to address this as users *must* have a domain, and the domain *must* be enabled for a user to authenticate. If a domain doesn’t exist for a user…it certainly can’t be enabled.

Short term, i’m removing the validation logic around the retrieving effective roles to account for users missing domains. This commit stops the tests from failing. A follow up commit will include tests to validate the condition is properly handled and returning the correct results.