Fixing failing noncommit failures due to pointcut proxies around class (#2074)
* Fixing DevOpsPurgeTrrIntegrationTest failure due to pointcut proxy around class

When a pointcut was added around the bean "tokenRevocationRecordPersistenceStrategy" it caused the tests in DevOpsPurgeTrrIntegrationTest to start failing due to spring creating a proxy around the bean. The bean in the spring container was therefore no longer a direct implementation of LdapTokenRevocationRecordRepository. The test class @Autowired a variable of type LdapTokenRevocationRecordRepository. Prior to the proxy this was fine, but after the proxy a bean of that type no longer existed in the spring context.

To work around switched the variable type to the interface. Given the test relies on the type being the explict LdapTokenRevocationRecordRepository, added a condition to skip the tests if the proxied bean is not of the correct type.

* Fixing Cloud20AEIntegrationTest test class containing concrete class reference