CID-1989: Add domainId attribute to auth request, clean up scopeaccess
To support the change we will need to store the domain id under which the user authenticated within a token. When looking at the
scope access object I noticed it still has all the unboundId annotations required for storing tokens in the directory. Since
we no longer store them in the directory these annotations are unnecessary, so I removed them.